Separate queue to component file
