+++ frontend/lyric-sync-app/src/api/apis/index.ts
@@ -4,3 +4,4 @@ export * from './HandwritingApi';
 export * from './QuizApi';
 export * from './SongApi';
 export * from './UserApi';
+export * from './AuthApi';
