+++ frontend/lyric-sync-app/src/components/QueueComponent/QueueComponent.scss
@@ -1,0 +1,35 @@
+@use '../../variables' as *;
+@use '../../mixins' as *;
+@use 'sass:color';
+
+.queue-column {
+  padding-right: 16px;
+  box-shadow: 2px 2px 6px rgba(0, 0, 0, 0.3);
+  border-radius: 8px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background: $gradient !important;
+  width: 100%;
+  height: 100%;
+  
+  .queue-title {
+    padding: 16px;
+    color: $text-primary;
+    font-weight: bold;
+  }
+  
+  .queue-list {
+    flex: 1;
+    padding: 0;
+    width: 100%;
+    overflow-y: scroll;
+    scrollbar-width: none; /* Firefox */
+    -ms-overflow-style: none;  /* Internet Explorer 10+ */
+  }
+}
+
+.queue-list::-webkit-scrollbar{ /* WebKit */
+  width: 0;
+  height: 0;
+}
